fix generic application definition
a bug in found in petshop sysl where `?` is only parsed as the beggining query params and the highlighter
expects a query params. This ruins scope definition as sometimes `?` can be used next to types to define
optional. This commit restricts it so that query params start with `?\w`.

More definitions is also added to the generic application definition scope for package and reference.