Fix incorrect SourceContext.file paths in protobuf modules (#55)
When calling command/transform plugins, the client first compiles the Sysl model to a protobuf message by passing the Sysl source to stdin of the Sysl binary. Due to the configuration of the parsing in versions < 0.468.0, this produces unexpected file paths in source context objects, which can be breaking for plugins that depend on those paths.

This issue is fixed in Sysl v0.468.0, so bumping the min Sysl version fixes this in the extension.