Support actions for LSP plugins (#59)
Add POC support for plugin **actions**: plugin-executed commands that can be listed and invoked from the client.

A plugin action is analogous to a command contributed by an extension, except that it is contributed dynamically by a plugin upon initialization. When the plugin sends its initialize response, it can include the capability `experimental.actions`, an array of actions that should be made available on the client. Each action must have at least an ID.

The extension contributes a `Sysl: List Actions` command that opens a quick pick menu from which to select an action. Upon selection, the extension invokes `executeCommand` with the ID of the selected action. It is assumed that the plugin that contributed the action will have registered to handle that command execution.

Other changes:
- Dynamically assign ports to LSP plugins to avoid collisions and allow debugging with multiple plugins.
- Automatically open (in memory and thus in plugins) `<name>.*.sysl` when `<name>.sysl` is opened. This is a temporary hack to make any auxiliary files available to the plugin servers through the regular text sync channel.
- Fix snapshots for multiviews.