Replace plugin clientSettings with async getter
The clientSettings property of Plugin appears to be available to plugin code, but since it is only set asynchronously after onInitialized(), there may be a race condition between it being set and code trying to use it, with no way to wait if it's unset.

To avoid this pitfall, the property is removed and replaced with an async function that fetches a fresh batch of settings from the client. This also avoids the cached settings going stale.