Compile and send Sysl models to plugins
The main purpose of this PR is to have the client compile Sysl files and send the compiled model to the plugins, rather than relying on each plugin compiling the model itself. This allows each plugin to work with the richer API of a TS-native Sysl model, and prevents duplicate compilation by each plugin.

This is to support the Sysl Plus plugin, which also needs a couple of other small new features (e.g. async validation and completion in the plugin template). As a side-effect, the ERD plugin has been uplifted to use the `plugin` package, and disposable/view types that were duplicated between `extension/views/types.ts` and `model`/`plugin` have been deduped.

Notes:

The current implementation is a little janky because the model change notifications are not in sync with the document changes (they are unaware of the model and don't wait for it to be compiled). It works well enough to expose the model though, and the performance and APIs can be iterated on.

There is still the possibility that a rapid succession of changes and saves causes many concurrent model compilations, which can be fatal on a large model. However compilations of the model are memoized, so hitting save twice for the same file content will only compile once. This mechanism can also be improved over time.

Both model and plugin need to be updated, so this may need to happen in a couple of steps. They have been updated to our more modern TS lib style though, so should be better to work with.