Add actions to plugin, rendering onModelChange
Add `Action` type to plugin to be shared by extension and plugins. (Backwards-incompatible change: renaming `action` property to `id`, because `action` was ambiguous. Extension can support both for a while.)

Also add an `onModelChange` callback to `Rendering` in the plugin template to support re-rendering when the model changes (probably the most common use case).