Fix publishing with Lerna (#93)
The previous refactor broke the publish workflows. This is an attempt to fix them.

The use of `lerna publish` (which uses `lerna version` behind the scenes) means:
- All packages that have changes will have their version bumped and published
- All packages that have their versions bumped will be bumped to the same version
- All packages will be published as npm packages, even if they don't really need to be (e.g. `packages/extension`, which is really only intended for publishing to VS Code marketplace).

Because the versions of the packages should/will be equal, I've bumped them all up to 0.54.0, which is a minor bump over the highest existing version (extension at 53).

It's very difficult to test, since we have to push to github.com/anz-bank/vscode-sysl first.