Improve robustness of download logic for Sysl/plugins (#107)
Replace the `download` package with a couple of custom `download` functions that do more or less the same thing. Using `got` directly improves the retry behaviour, and exposes more options for future consideration (e.g. DNS caching).

Also change the root URL for network plugins to an extension setting that defaults to the current hardcoded value. This allows users to work around the hardcoded value not working if necessary. If the value is empty, network plugins will be skipped with a log.