Pass arguments through to actions (#113)
When using the `List actions` Sysl extension quick pick, there is no facility for passing args to the action. However when invoking the action as a command (e.g. via a CodeLens supplied by a plugin) args can be provided, and if they are, they should be passed through to the plugins for handling.